fat32-fix: P0 生产加固 FAT32 驱动与 ATA 裸写 (A-F)
- fat_next: 坏簇/EOF 归一 EOF 哨兵, 非法簇号拦截(g_max_cluster)
- read_file_at: 预读不超过卷末 LBA, 修复大文件尾块越界丢数据
- fat32_mount: BPB 签名/类型/簇数校验, 记录 g_max_lba/g_max_cluster
- dir_lookup: LFN 校验和与 8.3 短名比对, 损坏 LFN 回退短名
- disk_write: 成功后使顺序读预读缓存失效
- ata.c: 写命令级重试(3次)+64位LBA(去 uint32_t 截断)
- fs_server self-test: 新增 6.8MB 大文件整读+尾块读校验

+++ include/kernel/ata.h
@@ -17,10 +17,10 @@
 bool ata_init(void);
 
 /* PIO 读 count 个扇区到 buf（须容纳 count*512 字节）；成功返回 true */
-bool ata_read_sectors(uint32_t lba, uint8_t count, void *buf);
+bool ata_read_sectors(uint64_t lba, uint8_t count, void *buf);
 
-/* PIO 写 count 个扇区（写后执行 FLUSH CACHE 确保落盘）；成功返回 true */
-bool ata_write_sectors(uint32_t lba, uint8_t count, const void *buf);
+/* PIO 写 count 个扇区（写后执行 FLUSH CACHE 确保落盘，命令级重试）；成功返回 true */
+bool ata_write_sectors(uint64_t lba, uint8_t count, const void *buf);
 
 /* 磁盘总扇区数（L2 修复：优先 LBA48 容量 word100-103，不支持时回退 LBA28
  * word60-61；>128GB 盘不再截断为 LBA28 上限） */
